# contracts/openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResuAlign API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/workbench/session/init": {
      "post": {
        "summary": "Init Workbench Session",
        "description": "Create a workstation session from a pasted JD or JD URL.\n\nReturns immediately with a WorkstationState; crawling, classification,\nand JD profiling run on a background worker and stream through SSE.",
        "operationId": "init_workbench_session_api_workbench_session_init_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkbenchSessionInitRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkstationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/session/{session_id}": {
      "get": {
        "summary": "Get Workbench Session",
        "description": "Return the current workstation state, with polling via If-None-Match.",
        "operationId": "get_workbench_session_api_workbench_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkstationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/session/{job_id}": {
      "get": {
        "summary": "Get Workspace Session",
        "description": "Open an existing library job as a read-only workstation session.",
        "operationId": "get_workspace_session_api_workspace_session__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkstationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/session/{session_id}/analyze": {
      "post": {
        "summary": "Analyze Workbench Session",
        "description": "Run JD profiling/gap analysis for a library-job workstation session.\n\nThe first call queues the pipeline; subsequent calls return the current\nstate without spawning another worker.",
        "operationId": "analyze_workbench_session_api_workbench_session__session_id__analyze_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkstationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/session/{session_id}/events": {
      "get": {
        "summary": "Stream Workbench Events",
        "description": "Stream workstation events as Server-Sent Events with history replay.",
        "operationId": "stream_workbench_events_api_workbench_session__session_id__events_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "replay",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Replay"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkbenchSessionInitRequest": {
        "properties": {
          "raw_jd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Jd"
          },
          "jd_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jd Url"
          },
          "master_resume_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Master Resume Id"
          },
          "resume_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Text"
          },
          "granularity": {
            "type": "string",
            "enum": [
              "fine",
              "medium",
              "coarse"
            ],
            "title": "Granularity",
            "default": "medium"
          },
          "prompt_focus": {
            "type": "string",
            "enum": [
              "balanced",
              "quantified",
              "skills"
            ],
            "title": "Prompt Focus",
            "default": "balanced"
          },
          "custom_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Prompt"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "title": "WorkbenchSessionInitRequest",
        "description": "Create a workstation session from a pasted JD or a JD URL."
      },
      "WorkstationAlignmentSection": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "default": "idle"
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "default": ""
          },
          "diffs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Diffs"
          },
          "invalid_diffs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Invalid Diffs"
          },
          "draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "eval_score": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Score"
          }
        },
        "type": "object",
        "title": "WorkstationAlignmentSection"
      },
      "WorkstationCrawlSection": {
        "properties": {
          "crawl_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crawl Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "queued",
              "fetching",
              "parsing",
              "classifying",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "default": "idle"
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "WorkstationCrawlSection"
      },
      "WorkstationGapSection": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "ready",
              "failed",
              "blocked"
            ],
            "title": "Status",
            "default": "queued"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "gap_report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Report"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "WorkstationGapSection"
      },
      "WorkstationJDSection": {
        "properties": {
          "profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "ready",
              "failed"
            ],
            "title": "Status",
            "default": "queued"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "WorkstationJDSection"
      },
      "WorkstationMeta": {
        "properties": {
          "etag": {
            "type": "string",
            "title": "Etag",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          },
          "event_url": {
            "type": "string",
            "title": "Event Url",
            "default": ""
          }
        },
        "type": "object",
        "title": "WorkstationMeta"
      },
      "WorkstationResumeSection": {
        "properties": {
          "selected_resume_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Resume Id"
          },
          "available_resumes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Available Resumes"
          },
          "content_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Ref"
          }
        },
        "type": "object",
        "title": "WorkstationResumeSection"
      },
      "WorkstationState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "initializing",
              "ready",
              "failed"
            ],
            "title": "Status",
            "default": "initializing"
          },
          "job": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job"
          },
          "jd": {
            "$ref": "#/components/schemas/WorkstationJDSection"
          },
          "resume": {
            "$ref": "#/components/schemas/WorkstationResumeSection"
          },
          "gap": {
            "$ref": "#/components/schemas/WorkstationGapSection"
          },
          "alignment": {
            "$ref": "#/components/schemas/WorkstationAlignmentSection"
          },
          "appraisal": {
            "additionalProperties": true,
            "type": "object",
            "title": "Appraisal"
          },
          "crawl": {
            "$ref": "#/components/schemas/WorkstationCrawlSection"
          },
          "meta": {
            "$ref": "#/components/schemas/WorkstationMeta"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "WorkstationState"
      }
    }
  }
}